check path is writeable

+++ src/Module/Util/AssetCache.php
@@ -35,7 +35,7 @@ use RegexIterator;
  */
 class AssetCache
 {
-    const cacheText = '_cached_by_ampache_';
+    const CACHETEXT = '_cached_by_ampache_';
 
     /**
      * This uses the MD5 hash of a file to create a unique cached version, to avoid 'just clear your browser cache' issues
@@ -48,10 +48,10 @@ class AssetCache
         $originalURL   = $url;
         $originalPath  = self::get_path($originalURL);
 
-        $cachedURL  = $originalArray['dirname'] . '/' . $originalArray['filename'] . self::cacheText . md5_file(self::get_path($originalURL)) . '.' . $originalArray['extension'];
+        $cachedURL  = $originalArray['dirname'] . '/' . $originalArray['filename'] . self:: . md5_file(self::get_path($originalURL)) . '.' . $originalArray['extension'];
         $cachedPath = self::get_path($cachedURL);
 
-        if (!file_exists($cachedPath)) {
+        if (!file_exists($cachedPath) && is_writeable($cachedPath)) {
             self::copy_file($originalPath);
         }
 
@@ -73,7 +73,7 @@ class AssetCache
         $pathArray = pathinfo($path);
 
         if (file_exists($path)) {
-            $cachedVersion = $pathArray['dirname'] . '/' . $pathArray['filename'] . self::cacheText . md5_file($path) . '.' . $pathArray['extension'];
+            $cachedVersion = $pathArray['dirname'] . '/' . $pathArray['filename'] . self::CACHETEXT . md5_file($path) . '.' . $pathArray['extension'];
             copy($path, $cachedVersion);
         }
     }
@@ -82,7 +82,7 @@ class AssetCache
     {
         $directory = new RecursiveDirectoryIterator(Core::get_server('DOCUMENT_ROOT'));
         $iterator  = new RecursiveIteratorIterator($directory);
-        $files     = new RegexIterator($iterator, '/.+' . self::cacheText . '.+/i', RecursiveRegexIterator::GET_MATCH);
+        $files     = new RegexIterator($iterator, '/.+' . self::CACHETEXT . '.+/i', RecursiveRegexIterator::GET_MATCH);
 
         foreach ($files as $file) {
             $file = implode("", $file);
